Added seeking to a position in an Avro file
The Avro file seeking uses both blocks and records to seek to the requested
position. Seeking with relatively large block sizes allows for faster random
access into an Avro file. Possible improvements to this mechanism would be to
introduce an index which holds information about the position in the Avro file.

The JSON values returned by avro_record_read are now proper JSON objects instead
of an array of objects with only one value. This makes processing the resulting
objects much easier and simpler.