Improved client notification and added Avro block size managemet
The clients are now notified when new data arrives so that they don't have to
poll for it constantly. When a data block is flushed to disk, the clients are
notified that it is available.

The data block size is now also configurable. The size of the data blocks is
measured in row events or transactions. The default value is 1000 for row events
and 50 for transactions. When the amount of processed rows or transactions
exceeds the limits, the block is flushed to disk. This is when the clients are
notified.