MXS-631, MXS-632: Cleaned up default value CMake files
The VERSION file now contains the contents of the set_maxscale_version macro.
The macro only set the version number so it can be converted to a "static" file.
The VERSION file was extremely outdated and was not used for anything. Moving
parts of macros.cmake into the VERSION file allows it to be used again.

The default values for the configuration were stored in macros.cmake inside
a macro that just sets the values to their defaults. These values are now in
the cmake/defaults.cmake file which is more easy to find compared to
macros.cmake.