Added sending of schemas and file rotation
The schemas are now sent before the actual data is streamed. Once the file is
read and the next file exists, the client will send its schema and start
streaming that file.

Parts of the mysql_binlog.c were moved into the avrorouter because the data
structures were too interconnected with avrorouter data structures. Now the
mysql_binlog file only contains functions which read packed data and convert
it to more manageable format.