Added Avro file indexing
The Avro files are indexed by GTID using SQLite3. Clients requesting records
from a GTID position retireve the data more quickly if the file is already
indexed. The index can also be viewed by users to query for data about
the converted Avro files and which GTIDs are in them.