Fixed event position updating
The position of the next event to be written was used as the position
of the current event. This caused the checks for the position of the current
safe event to fail and the non-transaction safe version was used.

This only happened with events that are not done inside a transaction i.e.
DDL statements.